// models/survey-selector.ts
export class SurveySelector {
    SurveySelectorId: number;
    SurveyId: number;
    EmployeeId: number;
    FirstName: string;
    LastName: string;
    Email: string;
    Hash: string;
    StartDate: string;
    EndDate: string;
    SurveySelectorStatusId: number;
    CompletedPageId: number;
    LastAccessedPageId: number;
    Language: string;
}

export class SurveySelectorIds {
    SurveySelectorId: number[];
}